requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
pandas
sentence-transformers
requests